common-requirements.txt
# Basic Packages
Django==1.11.12
psycopg2-binary==2.7.4
six==1.11.0
pytz==2018.4
django-localflavor==1.4.1


# Task packages
celery==4.1.0
django-celery-results==1.0.1
django-celery-beat==1.1.1


# Useful packages
pygeocodio==0.7.0
django-html5==1.0.0
django_sekizai==0.10.0
django-crispy-forms==1.7.2
django-smalluuid==1.1.1
unicodecsv==0.14.1
awesome-slugify==1.6.5

# Environ for management
django-environ==0.4.4


# NewRelic for monitoring
newrelic==3.2.0.91


# API-Related Packages
djangorestframework==3.8.2
django-filter==1.1
django-cors-headers==2.0.2


# AWS Packages
boto==2.48.0
boto3==1.7.4
-e git+https://github.com/nickcatal/django-sns-view.git@5c3531b3f5d77b0d10f9a7c5df5b9726d8e31d28#egg=django_sns_view
django-storages==1.6.6
Collectfast==0.6.2
django-s3-folder-storage==0.5


# Testing-Related Packages
mock==2.0.0
model-mommy==1.5.1
nose==1.3.7
django-nose==1.4.4
testfixtures==6.0.0


# Debugging and general utilities
django-clear-cache==0.3
django_debug_toolbar==1.9.1
ipython==5.5.0
